backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# LangChain stack (aligned on core 0.3.x)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.13,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-google-genai>=2.0.3,<3.0.0
langchain-openai>=0.2.0,<0.3.0
langgraph>=0.2.14,<0.3.0

faiss-cpu==1.8.0.post1
mwparserfromhell==0.6.6
tiktoken==0.7.0
bs4==0.0.2